refactor PreFromApiBase
